Add support for RAID-z style parity to AnyRAID.
There are a few changes needed for this, but none is major. Genericizing
the existing anyraid code, adding a concept of "width" to account for
data columns as well as parity, and adding some helper functions.  The
main work is mostly adding the interface to the vdev_raidz code in
vdev_anyraid.c. We need to redo some of the raidz map creation process
to correct the offsets and child IDs of the columns in the map. We also
need to implement a couple more of the vdev_ops functions that raidz
needs that mirrors don't, and handle both cases properly. Finally, we
add some anyraidz cases to the tests.

Sponsored-by: Eshtek, creators of HexOS
Sponsored-by: Klara, Inc.
Contributions-by: JT Pennington <jt.pennington@klarasystems.com>

+++ tests/zfs-tests/tests/functional/anyraid/anyraid_faildisk_write_replace_resilver.ksh
@@ -49,42 +49,42 @@ cleanup() {
 log_onexit cleanup
 
 
-# anymirror1
+for vdev in "anymirror1" "anyraidz1:1"; do
+	for replace_flags in '' '-s'; do
 
-for replace_flags in '' '-s'; do
+		[[ "$vdev" == "anyraidz1" && "replace_flags" == "-s" ]] && continue
+		log_must create_sparse_files "disk" 3 $DEVSIZE
+		log_must create_sparse_files "spare" 1 $DEVSIZE
+		log_must zpool create -O compress=off -f $TESTPOOL $vdev $disks
+		log_must zfs set primarycache=none $TESTPOOL
 
-	log_must create_sparse_files "disk" 3 $DEVSIZE
-	log_must create_sparse_files "spare" 1 $DEVSIZE
-	log_must zpool create -O compress=off -f $TESTPOOL anymirror1 $disks
-	log_must zfs set primarycache=none $TESTPOOL
+		# Write initial data
+		log_must file_write -o create -f /$TESTPOOL/file1.bin -b 1048576 -c 256 -d Z
 
-	# Write initial data
-	log_must file_write -o create -f /$TESTPOOL/file1.bin -b 1048576 -c 256 -d Z
+		# Fail one disk
+		log_must truncate -s0 $disk0
 
-	# Fail one disk
-	log_must truncate -s0 $disk0
+		# Read initial data, write new data
+		log_must dd if=/$TESTPOOL/file1.bin of=/dev/null bs=1M count=256
+		log_must file_write -o create -f /$TESTPOOL/file1.bin -b 1048576 -c 256 -d Y
 
-	# Read initial data, write new data
-	log_must dd if=/$TESTPOOL/file1.bin of=/dev/null bs=1M count=256
-	log_must file_write -o create -f /$TESTPOOL/file1.bin -b 1048576 -c 256 -d Y
+		# Check that disk is faulted
+		zpool status
+		log_must check_state $TESTPOOL $disk0 "faulted"
 
-	# Check that disk is faulted
-	zpool status
-	log_must check_state $TESTPOOL $disk0 "faulted"
+		# Initiate disk replacement
+		log_must zpool replace -f $replace_flags $TESTPOOL $disk0 $spare0
 
-	# Initiate disk replacement
-	log_must zpool replace -f $replace_flags $TESTPOOL $disk0 $spare0
+		# Wait until resilvering is done and the pool is back online
+		for i in {1..60}; do
+			check_state $TESTPOOL "" "online" && break
+			sleep 1
+		done
+		zpool status
+		log_must check_state $TESTPOOL "" "online"
 
-	# Wait until resilvering is done and the pool is back online
-	for i in {1..60}; do
-		check_state $TESTPOOL "" "online" && break
-		sleep 1
+		destroy_pool $TESTPOOL
 	done
-	zpool status
-	log_must check_state $TESTPOOL "" "online"
-
-	destroy_pool $TESTPOOL
-
 done
 
 log_pass "AnyRAID mirror can resilver a replaced disk"
